Added Ceph to CI scripts and fixed problems found while exercising the Ceph code; however, the calls to parted from ceph_volume.yaml are failing as the volume on the master '/dev/vdb' is already in use.
Some of the problems stemmed from the assumption that the root user
would be kicking off the iaas_launch.py script which is no longer a
requirement. Only a passwordless sudo user with keyed access to the
nodes is required of the user starting the process.

Also observed logic issues around host and ceph volume claims. When more
than one of each is configured, only the second gets created.
Additionally, files from kubespray are getting altered which will break
other projects being deployed to the same rack.